+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.3] - 2022-10-23
 
 - Run destroy_all in batch
+- Override inspect method 
 
 ## [0.0.2] - 2022-10-23